requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.12.3
    # via bs4
blis==1.1.0
    # via thinc
bs4==0.0.2
    # via -r requirements.in
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
cloudpathlib==0.20.0
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cymem==2.0.10
    # via
    #   preshed
    #   spacy
    #   thinc
discordwebhook==1.0.3
    # via -r requirements.in
idna==3.10
    # via requests
jinja2==3.1.5
    # via spacy
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
lxml==5.3.0
    # via -r requirements.in
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==2.2.1
    # via
    #   blis
    #   spacy
    #   thinc
packaging==24.2
    # via
    #   spacy
    #   thinc
    #   weasel
praw==7.8.1
    # via -r requirements.in
prawcore==2.4.0
    # via praw
preshed==3.0.9
    # via
    #   spacy
    #   thinc
pydantic==2.10.5
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   discordwebhook
    #   prawcore
    #   spacy
    #   update-checker
    #   weasel
rich==13.9.4
    # via typer
setuptools==75.8.0
    # via
    #   marisa-trie
    #   spacy
    #   thinc
shellingham==1.5.4
    # via typer
smart-open==7.1.0
    # via weasel
soupsieve==2.6
    # via beautifulsoup4
spacy==3.8.3
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.0
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
thinc==8.3.3
    # via spacy
tqdm==4.67.1
    # via spacy
typer==0.15.1
    # via
    #   spacy
    #   weasel
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   typer
update-checker==0.18.0
    # via praw
urllib3==2.3.0
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
websocket-client==1.8.0
    # via praw
wrapt==1.17.0
    # via smart-open
